Move all test classes to a sibling package
When test classes live in the same package, where auto-configuration classes
under test also live, these tests are not pure anymore because they are under
the same component scan, as main classes. This component scan could create
beans and other entities that are not supposed to be created and will not be
created in a real application that will use problem-spring-web-autoconfigure
module.

This commit moves all test classes into a sibling package, so component scan
will not find any components from the main code, and container will contain
only beans created by auto-configurations.