Fix registration of SecurityExceptionHandling controller advice
When bean-factory method returns `AdviceTrait` interface that is not directly
annotated with `@ControllerAdvice`, Spring doesn't register this bean
as controller advice, so all `@ExceptionHander`s it provides are also
not registered.

This happens probably because of ordering between beans instantiation and
registration of controller advices in Spring MVC.

This commit changes the return type of bean-factory methods to a more concrete
type, that is directly annotated with `@ControllerAdvice` annotation.
This makes Spring always find the `@ControllerAdvice` annotation and
properly register these controller advices and any exception handlers
they contain.