Improve testing of configuration of exception handling
Previously the test only checked that ExceptionHandling bean is present
in the context. But the presense of the bean doesn't mean the controller advice
and associated exception handlers are registered in Spring MVC.

Now the test will also check if HTTP request is really handled by one of
problem-spring-web exception handlers.